Refactor movement handling and player interactions (Delays etc)
Replaced `MovementHandler` with `PlayerMovementHandler` for improved specificity and handling of player actions. Introduced `InteractingEntity` and `MaxDistance` abstractions in `RSInteraction` to standardize interaction targets and distances. Updated interaction logic for activities like mining, banking, and combat to align with the new handler and interaction flow.

+++ Genesis/Movement/MovementHelper.cs
@@ -72,7 +72,7 @@ public static class MovementHelper
         return (DIRECTION_DELTA_X[(int)direction], DIRECTION_DELTA_Y[(int)direction]);
     }
     
-    public static double EuclideanDistance(double x1, double y1, double x2, double y2)
+    public static double EuclideanDistance(int x1, int y1, int x2, int y2)
     {
         return Math.Sqrt(Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2));
     }
